#lightbar: InputReactive flashes on the Mic/Share button and touchpad click
The InputReactive rising-edge detector read only the 16-bit Gamepad.Buttons mask,
which is full (DPAD..Y). Two presses live outside it and so never flashed:

- The Share / Create button — its own bool field on the Gamepad struct. The physical
  Mic / Misc1 button maps to ButtonShare (control index 11), which sets gp.Share, so
  pressing it produced no change in the mask the detector watched.
- The touchpad click — folded here from the physical TouchpadOutputState so it fires
  even if the click is remapped away from the TOUCHPAD output bit.

Widened SlotButtonsProvider + the detector to uint and fold both presses onto spare
bits (Share -> 0x10000, touchpad click -> 0x20000).

+++ PadForge.App/Common/Input/UserEffectsDispatcher.cs
@@ -102,7 +102,7 @@ namespace PadForge.Common.Input
         /// <c>InputManager.CombinedOutputStates[i].Buttons</c>. Used by
         /// <see cref="LightbarMode.InputReactive"/> to detect rising edges
         /// and enqueue a fading pulse.</summary>
-        public static Func<int, ushort> SlotButtonsProvider { get; set; }
+        public static Func<int, uint> SlotButtonsProvider { get; set; }
 
         /// <summary>Static provider for the current rumble state of a
         /// given (slot, physical device) pair, returned as 8-bit
@@ -205,7 +205,7 @@ namespace PadForge.Common.Input
         private uint _randomColor;
         private bool _audioOnsetActive;
         private long _pulseStartMs;
-        private ushort _lastButtons;
+        private uint _lastButtons;
         private readonly Random _rng = new Random();
 
         private sealed class DeviceState
@@ -893,8 +893,8 @@ namespace PadForge.Common.Input
             // Slot-level button-press detection — one rising-edge event
             // per tick fans out to every per-device pulse below.
             var provider = SlotButtonsProvider;
-            ushort buttons = provider != null ? provider(_padIndex) : (ushort)0;
-            ushort newlyPressed = (ushort)(buttons & ~_lastButtons);
+            uint buttons = provider != null ? provider(_padIndex) : 0u;
+            uint newlyPressed = buttons & ~_lastButtons;
             _lastButtons = buttons;
             if (newlyPressed == 0) return;
